moved tests to proper package
