+++ src/test/java/io/zipcoder/casino/Models/DieTest.java
@@ -1,12 +1,20 @@
 package io.zipcoder.casino.Models;
 
+import io.zipcoder.casino.Models.Die;
+import org.junit.Assert;
 import org.junit.Test;
-
-import static org.junit.Assert.*;
 
 public class DieTest {
 
     @Test
-    public void getDiceRoll() {
+    public void getDiceRollTest() {
+        // Given
+        Die die = new Die();
+
+        // When
+        Integer result = die.getSingleDieRoll();
+
+        // Then
+        Assert.assertTrue(result >= 1 && result <= 6);
     }
 }
